+++ cf/data/array/mixin/arraymixin.py
@@ -9,10 +9,18 @@ class ArrayMixin:
     """
 
     def __array_function__(self, func, types, args, kwargs):
-        """Implement the `numpy` ``__array_function__`` protocol."""
+        """Implement the `numpy` ``__array_function__`` protocol.
+
+        .. versionadded:: TODODASKVER
+
+        """
         return NotImplemented
 
     @property
     def Units(self):
-        """TODODASKDOCS."""
+        """The `cf.Units` object containing the units of the array.
+
+        .. versionadded:: TODODASKVER
+
+        """
         return Units(self.get_units(), self.get_calendar(None))