+++ cf/data/mixin/deprecations.py
@@ -316,6 +316,9 @@ class DataClassDeprecationsMixin:
     def chunk(self, chunksize=None, total=None, omit_axes=None, pmshape=None):
         """Partition the data array.
 
+        Deprecated at version TODODASKVER. Use the `rechunk` method
+        instead.
+
         :Parameters:
 
             chunksize: `int`, optional
@@ -341,7 +344,11 @@ class DataClassDeprecationsMixin:
 
         """
         _DEPRECATION_ERROR_METHOD(
-            "TODODASKMSG. Use 'rechunk' instead"
+            self,
+            "chunk",
+            message="Use the 'rechunk' method instead.",
+            version="TODODASKVER",
+            removed_at="5.0.0",
         )  # pragma: no cover
 
     def dumpd(self):
@@ -586,6 +593,9 @@ class DataClassDeprecationsMixin:
     def add_partitions(self, extra_boundaries, pdim):
         """Add partition boundaries.
 
+        Deprecated at version TODODASKVER. Use the `rechunk` method
+        instead.
+
         :Parameters:
 
             extra_boundaries: `list` of `int`
@@ -604,7 +614,11 @@ class DataClassDeprecationsMixin:
 
         """
         _DEPRECATION_ERROR_METHOD(
-            "TODODASKMSG Consider using rechunk instead"
+            self,
+            "add_partitions",
+            message="Use the 'rechunk' method instead.",
+            version="TODODASKVER",
+            removed_at="5.0.0",
         )  # pragma: no cover
 
     @staticmethod
@@ -727,6 +741,9 @@ class DataClassDeprecationsMixin:
         """Return the partition boundaries for each partition matrix
         dimension.
 
+        Deprecated at version TODODASKVER. Consider using the `chunks`
+        attribute instead.
+
         :Returns:
 
             `dict`
@@ -735,7 +752,11 @@ class DataClassDeprecationsMixin:
 
         """
         _DEPRECATION_ERROR_METHOD(
-            "TODODASKMSG - consider using 'chunks' instead"
+            self,
+            "partition_boundaries",
+            message="Consider using the 'chunks' attribute instead.",
+            version="TODODASKVER",
+            removed_at="5.0.0",
         )  # pragma: no cover
 
     def save_to_disk(self, itemsize=None):