+++ cf/mixin/propertiesdatabounds.py
@@ -1298,14 +1298,18 @@ class PropertiesDataBounds(PropertiesData):
         )  # pragma: no cover
 
     @classmethod
-    def concatenate(cls, variables, axis=0, _preserve=True):
+    def concatenate(cls, variables, axis=0, cull_graph=True):
         """Join a sequence of variables together.
+
+        .. seealso:: `Data.cull_graph`
 
         :Parameters:
 
             variables: sequence of constructs
 
             axis: `int`, optional
+
+            {{cull_graph: `bool`, optional}}
 
         :Returns:
 
@@ -1317,14 +1321,14 @@ class PropertiesDataBounds(PropertiesData):
         if len(variables) == 1:
             return variable0.copy()
 
-        out = super().concatenate(variables, axis=axis, _preserve=_preserve)
+        out = super().concatenate(variables, axis=axis, cull_graph=cull_graph)
 
         bounds = variable0.get_bounds(None)
         if bounds is not None:
             bounds = bounds.concatenate(
                 [v.get_bounds() for v in variables],
                 axis=axis,
-                _preserve=_preserve,
+                cull_graph=cull_graph,
             )
             out.set_bounds(bounds, copy=False)
 
@@ -1333,7 +1337,7 @@ class PropertiesDataBounds(PropertiesData):
             interior_ring = interior_ring.concatenate(
                 [v.get_interior_ring() for v in variables],
                 axis=axis,
-                _preserve=_preserve,
+                cull_graph=cull_graph,
             )
             out.set_interior_ring(interior_ring, copy=False)